Moves classes out of sample app
